backend/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# CLARION — Python dependencies
# Install: pip install -r requirements.txt (prefer: python -m pip install -r requirements.txt)
# ─────────────────────────────────────────────────────────────────────────────

# Core
pydantic>=2.0.0
python-dotenv>=1.0.0

# API server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
# pymupdf
# pdfplumber
google-cloud-firestore>=2.20.0
google-cloud-storage>=2.16.0
google-cloud-tasks>=2.18.0
google-genai>=1.66.0
google-adk>=1.26.0
pdfplumber>=0.10.0
